Create reusable Wrapper component for handling inline logic
